requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
supabase==2.5.3
openai==1.35.13
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.0
